chore: update project version to 1.3.1 and enhance description in pyproject.toml; add shared testing fixtures and improve CodeQL workflow
