+++ backend/pyproject.toml
@@ -4,8 +4,13 @@ build-backend = "hatchling.build"
 
 [project]
 name = "django-crm"
-version = "1.3.0"
-description = "Open-source multi-tenant CRM platform built on Django REST Framework with row-level security."
+version = "1.3.1"
+# The one-line summary PyPI shows under the project title and in search results,
+# so it carries the terms people actually search for ("open source CRM",
+# "self-hosted CRM", "Django CRM") and names the modules rather than the
+# architecture. Published release metadata is immutable, so a change here only
+# reaches PyPI on the next version bump.
+description = "Open-source self-hosted CRM with a Django REST API: leads, deals, contacts, invoices and support tickets, multi-tenant by PostgreSQL row-level security."
 readme = "README.md"
 requires-python = ">=3.12"
 license = { text = "MIT" }
@@ -23,6 +28,19 @@ keywords = [
     "leads",
     "contacts",
     "opportunity",
+    # Added for discovery: these are the phrases people search rather than the
+    # words the codebase happens to use. "self-hosted" and "open-source-crm"
+    # are how someone shopping for an alternative describes what they want.
+    "self-hosted",
+    "open-source-crm",
+    "sales-pipeline",
+    "invoicing",
+    "helpdesk",
+    "support-tickets",
+    "rest-api",
+    "postgresql",
+    "row-level-security",
+    "salesforce-alternative",
 ]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -78,11 +96,16 @@ dependencies = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/django-crm/Django-CRM"
-Repository = "https://github.com/django-crm/Django-CRM"
-Documentation = "https://github.com/django-crm/Django-CRM#readme"
-Issues = "https://github.com/django-crm/Django-CRM/issues"
-Changelog = "https://github.com/django-crm/Django-CRM/releases"
+# Owner spelled `Django-CRM`, matching GitHub's own casing. The lowercase form
+# these used to carry resolves fine, because GitHub URLs are case-insensitive,
+# which is exactly what makes the habit dangerous: the same lowercase spelling
+# in the PyPI trusted publisher silently matched nothing and cost three failed
+# releases of 1.3.0. Keep every reference to the repo in GitHub's casing.
+Homepage = "https://github.com/Django-CRM/Django-CRM"
+Repository = "https://github.com/Django-CRM/Django-CRM"
+Documentation = "https://github.com/Django-CRM/Django-CRM#readme"
+Issues = "https://github.com/Django-CRM/Django-CRM/issues"
+Changelog = "https://github.com/Django-CRM/Django-CRM/releases"
 
 [dependency-groups]
 dev = [
